requirements.txt
asgiref==3.9.1
Django==5.2.4
django-extensions==4.1
django-js-asset==3.1.2
django-mptt==0.17.0
django-mptt-admin==2.8.0
django-nested-admin==4.1.1
djangorestframework==3.16.0
iniconfig==2.1.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
python-monkey-business==1.1.0
sqlparse==0.5.3
